Work around issue with ember-source < 3.1.0.
Ember < 3.1.0 mutates `element.tag` so that the first character
is _always_ lower case (even if the raw source includes upper case).

This commit adds a helper function that grabs the first char from
the raw source when possible to detect the _actual_ first char (falling
back to whatever `element.tag` was if contents are not provided).