Fix transform to not touch native <input>
Fixes #78

That was a pretty stupid one: the transform would treat `<input />` the same as `<Input />`, i.e. transform it to `{{input}}`. 🤦‍♂️
Fixed that, added a test, and also tested it with the test suite of ember-bootstrap, which uncovered this issue as reported in #78.

Had to extract some of the existing helpers that work around the issue of earlier Ember version's AST returning the lower-cased tag name, tu reuse them here.