Yarn upgrade
This fixes a DX issue: it seems there was some change in prettier between the currently locked version and the latest one. This leads to a situation where `yarn lint:js` would pass locally (when deps have been installed with the lockfile), but fails on CI (where `yarn --no-lockfile` is used).